edit Carousel: convert to tsx
